+++ opm/simulators/linalg/system/SystemPreconditioner.hpp
@@ -30,16 +30,6 @@
 
 namespace Opm {
 
-// Reservoir operator/comm types used as template arguments.
-template<typename Scalar>
-using SeqResOperator = Dune::MatrixAdapter<RRMatrix<Scalar>, ResVector<Scalar>, ResVector<Scalar>>;
-
-#if HAVE_MPI
-using ParResComm = Dune::OwnerOverlapCopyCommunication<int, int>;
-template<typename Scalar>
-using ParResOperator = Dune::OverlappingSchwarzOperator<RRMatrix<Scalar>, ResVector<Scalar>, ResVector<Scalar>, ParResComm>;
-#endif
-
 // Preconditioner for the coupled reservoir-well system.
 //
 // Templated on scalar type, reservoir operator and communication types to
@@ -67,29 +57,14 @@ public:
     SystemPreconditioner(const SystemMatrix<Scalar>& S,
                          const std::function<ResVector<Scalar>()>& weightsCalculator,
                          int pressureIndex,
-                         const Opm::PropertyTree& prm)
-        requires (!isParallel)
-        : S_(S)
-        , pressureIndex_(pressureIndex)
-    {
-        initSubSolvers(prm, weightsCalculator);
-        initWorkVectors();
-    }
+                         const Opm::PropertyTree& prm) requires (!isParallel);
 
     // Parallel constructor (enabled only for parallel specializations).
     SystemPreconditioner(const SystemMatrix<Scalar>& S,
                          const std::function<ResVector<Scalar>()>& weightsCalculator,
                          int pressureIndex,
                          const Opm::PropertyTree& prm,
-                         const ResComm& resComm)
-        requires (isParallel)
-        : S_(S)
-        , resComm_(&resComm)
-        , pressureIndex_(pressureIndex)
-    {
-        initSubSolvers(prm, weightsCalculator);
-        initWorkVectors();
-    }
+                         const ResComm& resComm) requires (isParallel);
 
     void pre(SystemVector<Scalar>&, SystemVector<Scalar>&) override
     {
@@ -107,20 +82,9 @@ public:
             return Dune::SolverCategory::sequential;
     }
 
-    void update() override
-    {
-        resSolver_->preconditioner().update();
-        resSmoother_->preconditioner().update();
-        wellSolver_->preconditioner().update();
-    }
+    void update() override;
 
-    void updateForChangedWellStructure()
-    {
-        resSolver_->preconditioner().update();
-        resSmoother_->preconditioner().update();
-        initWellSolver();
-        resizeWellWorkVectors();
-    }
+    void updateForChangedWellStructure();
 
     bool hasPerfectUpdate() const override
     {
@@ -136,68 +100,7 @@ public:
 //   C = reservoir-well coupling (top-right)
 //   B = well-reservoir coupling (bottom-left)
 //   D = well-well (bottom-right)
-     void apply(SystemVector<Scalar>& v, const SystemVector<Scalar>& d) override
-    {
-        // Extract blocks using the agreed convention
-        const auto& A = S_[_0][_0];
-        const auto& C = S_[_0][_1];
-        const auto& B = S_[_1][_0];
-        const auto& D = S_[_1][_1];
-
-        resRes_ = d[_0];
-        wRes_ = d[_1];
-        resSol_ = 0.0;
-        wSol_ = 0.0;
-
-        // Stage 1: Reservoir CPR solve
-        {
-            Dune::InverseOperatorResult res_result;
-            dresSol_ = 0.0;
-            tmp_resRes_ = resRes_;
-            syncResVector(tmp_resRes_);
-            resSolver_->apply(dresSol_, tmp_resRes_, res_result);
-            resSol_ += dresSol_;
-            // resRes_ -= A * dresSol_
-            A.mmv(dresSol_, resRes_);
-            // wRes_ -= B * dresSol_
-            B.mmv(dresSol_, wRes_);
-        }
-
-        // Stage 2: Well solve + reservoir system smoothing
-        {
-            Dune::InverseOperatorResult well_result;
-            dwSol_ = 0.0;
-            tmp_wRes_ = wRes_;
-            wellSolver_->apply(dwSol_, tmp_wRes_, well_result);
-            wSol_ += dwSol_;
-            // resRes_ -= C * dwSol_
-            C.mmv(dwSol_, resRes_);
-            // resRes_ -= D * dwSol_
-            D.mmv(dwSol_, wRes_);
-
-            Dune::InverseOperatorResult res_result;
-            dresSol_ = 0.0;
-            tmp_resRes_ = resRes_;
-            syncResVector(tmp_resRes_);
-            resSmoother_->apply(dresSol_, tmp_resRes_, res_result);
-            resSol_ += dresSol_;
-            // wRes_ -= B * dresSol_
-            B.mmv(dresSol_, wRes_);
-        }
-
-        // Stage 3: Final well solve
-        {
-            Dune::InverseOperatorResult well_result;
-            dwSol_ = 0.0;
-            tmp_wRes_ = wRes_;
-            wellSolver_->apply(dwSol_, tmp_wRes_, well_result);
-            wSol_ += dwSol_;
-        }
-
-        syncResVector(resSol_);
-        v[_0] = resSol_;
-        v[_1] = wSol_;
-    }
+    void apply(SystemVector<Scalar>& v, const SystemVector<Scalar>& d) override;
 
 private:
     const SystemMatrix<Scalar>& S_;
@@ -221,68 +124,18 @@ private:
     ResVector<Scalar> resRes_;
     WellVector<Scalar> wRes_;
 
-    void syncResVector(ResVector<Scalar>& v)
-    {
-        if constexpr (isParallel) {
-            resComm_->copyOwnerToAll(v, v);
-        }
-    }
+    void syncResVector(ResVector<Scalar>& v);
 
-    void initWellSolver()
-    {
-        wop_ = std::make_unique<WellOperator>(S_[_1][_1]);
-        std::function<WellVector<Scalar>()> weightsCalculatorWell;
-        wellSolver_ = std::make_unique<WellFlexibleSolverType>(
-            *wop_, wellprm_, weightsCalculatorWell, dummyWellPressureIndex);
-    }
+    void initWellSolver();
 
     void initSubSolvers(const Opm::PropertyTree& prm,
-                        const std::function<ResVector<Scalar>()>& weightsCalculator)
-    {
-        auto resprm = prm.get_child("reservoir_solver");
-        auto resprmsmoother = prm.get_child("reservoir_smoother");
-        wellprm_ = prm.get_child("well_solver");
+                        const std::function<ResVector<Scalar>()>& weightsCalculator);
 
-        if constexpr (isParallel) {
-            rop_ = std::make_unique<ResOp>(S_[_0][_0], *resComm_);
-            resSolver_ = std::make_unique<ResFlexibleSolverType>(
-                *rop_, *resComm_, resprm, weightsCalculator, pressureIndex_);
-            resSmoother_ = std::make_unique<ResFlexibleSolverType>(
-                *rop_, *resComm_, resprmsmoother, weightsCalculator, pressureIndex_);
-        } else {
-            rop_ = std::make_unique<ResOp>(S_[_0][_0]);
-            resSolver_ = std::make_unique<ResFlexibleSolverType>(
-                *rop_, resprm, weightsCalculator, pressureIndex_);
-            resSmoother_ = std::make_unique<ResFlexibleSolverType>(
-                *rop_, resprmsmoother, weightsCalculator, pressureIndex_);
-        }
+    void initWorkVectors();
 
-        initWellSolver();
-    }
+    void resizeReservoirWorkVectors();
 
-    void initWorkVectors()
-    {
-        resizeReservoirWorkVectors();
-        resizeWellWorkVectors();
-    }
-
-    void resizeReservoirWorkVectors()
-    {
-        const auto numRes = S_[_0][_0].N();
-        resSol_.resize(numRes);
-        dresSol_.resize(numRes);
-        tmp_resRes_.resize(numRes);
-        resRes_.resize(numRes);
-    }
-
-    void resizeWellWorkVectors()
-    {
-        const auto numWell = S_[_1][_1].N();
-        wSol_.resize(numWell);
-        dwSol_.resize(numWell);
-        tmp_wRes_.resize(numWell);
-        wRes_.resize(numWell);
-    }
+    void resizeWellWorkVectors();
 };
 
 } // namespace Opm
